Add vuex-utils for reading jsonApi state objects
